Complete medicationlist screen

+++ app/src/main/java/es/usj/mastertsa/cuidameapp/ui/indication/list/IndicationListScreen.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 
 @Composable
 fun IndicationListScreen(
-    navigateToDetail:(id: Int) -> Unit
+    navigateToDetail:(id: Long) -> Unit
 ){
 
 }
